requirements.txt
discord.py
datetime
requests